{"schema":"decodex/commit/1","summary":"Split quantitative product export rows","authority":"manual"}
