+++ apps/elf-eval/src/bin/real_world_job_benchmark/quantitative/product_manifest/export/rows.rs
@@ -1,55 +1,4 @@
-use crate::{QuantitativeBenchmarkRow, QuantitativePerQueryRow, RealWorldReport};
+mod per_query;
+mod product;
 
-pub(super) fn exported_product_row(
-	source_row: &QuantitativeBenchmarkRow,
-	product: &str,
-	adapter_id: &str,
-	adapter_name: &str,
-) -> QuantitativeBenchmarkRow {
-	let mut row = source_row.clone();
-
-	row.product = product.to_string();
-	row.adapter_id = adapter_id.to_string();
-	row.adapter_name = adapter_name.to_string();
-	row.claim_boundary = concat!(
-		"Exported from a generated real_world_job_report quantitative row; ",
-		"import remains subject to same-corpus, per-query, explicit-qrel, and leaderboard gates."
-	)
-	.to_string();
-
-	row
-}
-
-pub(super) fn exported_per_query_rows(
-	report: &RealWorldReport,
-	source_product: &str,
-	source_adapter_id: &str,
-	product: &str,
-	adapter_id: &str,
-) -> Vec<QuantitativePerQueryRow> {
-	report
-		.quantitative_scoreboard
-		.per_query_rows
-		.iter()
-		.filter(|row| row.product == source_product && row.adapter_id == source_adapter_id)
-		.map(|row| exported_per_query_row(row, product, adapter_id))
-		.collect()
-}
-
-fn exported_per_query_row(
-	source_row: &QuantitativePerQueryRow,
-	product: &str,
-	adapter_id: &str,
-) -> QuantitativePerQueryRow {
-	let mut row = source_row.clone();
-
-	row.product = product.to_string();
-	row.adapter_id = adapter_id.to_string();
-	row.claim_boundary = concat!(
-		"Exported from generated report per-query quantitative evidence; ",
-		"import does not relax paired-significance or leaderboard gates."
-	)
-	.to_string();
-
-	row
-}
+pub(super) use self::{per_query::exported_per_query_rows, product::exported_product_row};
